# skills/kanban-openapi.yaml
openapi: 3.0.0
info:
  title: AI Marketing Crew Agent API
  description: API for AI Agents to interact with the AI Marketing Crew collaboration board. Enables task creation, state transition, human-in-the-loop requests, and task retrieval.
  version: 1.0.0
servers:
  - url: https://amc-kanban.immedi.ai/api
    description: Production Server
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /agent/brand-config:
    get:
      summary: List or read brands linked to the current Agent
      operationId: getAgentBrandConfig
      description: Startup call. Omit brandId to list all active subscription-backed brands linked to this Agent, then record each id/name as KANBAN_BRAND_IDS and KANBAN_BRAND_LIST. Pass brandId to read one brand config; the server verifies the current Agent has an active BrandAgent link.
      parameters:
        - name: brandId
          in: query
          required: false
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Linked brand list or one brand config
    patch:
      summary: Update linked brand profile/config
      operationId: updateAgentBrandConfig
      description: Update an existing brand linked to this Agent. brandId is required and the server double-checks the current Agent has an active BrandAgent link before any mutation. Brand creation is disabled for Agents and must happen through dashboard subscription flow.
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brandId]
              properties:
                brandId:
                  type: string
                name:
                  type: string
                description:
                  type: string
                logoUrl:
                  type: string
                website:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                location:
                  type: string
                timezone:
                  type: string
                postfastApiKey:
                  type: string
                  description: Sensitive. Prefer human configuration in dashboard when possible.
                googlePlaceId:
                  type: string
                googleApiKey:
                  type: string
                  description: Sensitive. Prefer human configuration in dashboard when possible.
                googleAccountId:
                  type: string
                googleLocationId:
                  type: string
                googleLocationName:
                  type: string
                larkAppId:
                  type: string
                larkAppSecret:
                  type: string
                  description: Sensitive. Prefer human configuration in dashboard when possible.
                larkDriveFolderId:
                  type: string
                larkBotWebhook:
                  type: string
                larkOwnerId:
                  type: string
      responses:
        '200':
          description: Brand config updated
    post:
      summary: Disabled brand creation endpoint
      operationId: createBrandConfigDisabled
      description: Agents must not create brands. Returns 402 SUBSCRIPTION_REQUIRED_BEFORE_BRAND_CREATE.
      responses:
        '402':
          description: Brand creation requires dashboard subscription flow
...
  /brands/{brandId}/accounts:
    get:
      summary: List brand social accounts
      operationId: listBrandSocialAccounts
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Social account list
    post:
      summary: Add a brand social account record
      operationId: createBrandSocialAccount
      parameters:
        - name: brandId
          in: path
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SocialAccountInput'
      responses:
        '201':
          description: Social account created
...
components:
  schemas:
...
    SocialAccountInput:
      type: object
      required:
        - platformId
        - handle
      properties:
        platformId:
          type: string
          description: instagram, tiktok, google, facebook, yelp, etc.
        handle:
          type: string
        displayName:
          type: string
        profileUrl:
          type: string
        loginUsername:
          type: string
        loginPassword:
          type: string
          description: Sensitive; dashboard masks this for non-admin users.
...